inference-service/requirements.txt
openai==1.30.0
httpx==0.27.0
redis==5.0.0
python-dotenv==1.0.0
fastapi==0.111.0
uvicorn==0.29.0
prometheus-client==0.20.0